Created cards.js file
